refactoring fee to a file
